refactor(transitions): split parity presets into theme files and drop nested ternaries
