+++ qcut/apps/web/src/components/editor/media-panel/views/transitions/transition-category-expansions/jianying-parity.ts
@@ -1,11 +1,14 @@
 import type { TransitionPreset } from "../transition-preset-types";
 import { categoryExpansion } from "./build-category-expansion";
+import { JIANYING_EXTRA_PARITY_EXPANSIONS } from "./jianying-parity-extras";
+import { JIANYING_MASK_PARITY_EXPANSIONS } from "./jianying-parity-masks";
 
 /**
  * JianYing-parity parameter sweep: named looks from 剪映's 热门/叠化/运镜/
  * 光效/故障/综艺/模糊/幻灯片/自然/拍摄 shelves, expressed entirely through
- * the existing 20 clip transition types with tuned direction/intensity/
- * frequency/tint — no new render pipeline.
+ * the existing clip transition engines with tuned direction/intensity/
+ * frequency/tint. Mask-shape and later-round groups live in
+ * jianying-parity-masks.ts and jianying-parity-extras.ts.
  */
 const dissolveParity = categoryExpansion({
 	category: "dissolve",
@@ -414,738 +417,6 @@ const shootingParity = categoryExpansion({
 	],
 });
 
-const maskParityDissolve = categoryExpansion({
-	category: "dissolve",
-	rows: [
-		[
-			"circle-expand",
-			"Circle Expand",
-			"圆圈扩散",
-			"texture",
-			"texture-mask",
-			0.6,
-			{ maskShape: "circle", tags: ["circle", "mask"], popular: true },
-		],
-	],
-});
-
-const maskParitySlideshow = categoryExpansion({
-	category: "slideshow",
-	rows: [
-		[
-			"clock-wipe",
-			"Clock Wipe",
-			"旋转擦除",
-			"texture",
-			"texture-mask",
-			0.7,
-			{ maskShape: "clock", tags: ["clock", "sweep"], popular: true },
-		],
-		[
-			"blinds-sweep",
-			"Blinds Sweep",
-			"百叶窗",
-			"texture",
-			"texture-mask",
-			0.6,
-			{ maskShape: "blinds", tags: ["blinds"] },
-		],
-		[
-			"arrow-sweep",
-			"Arrow Sweep",
-			"箭头右扫",
-			"texture",
-			"texture-mask",
-			0.55,
-			{ maskShape: "arrow", tags: ["arrow"] },
-		],
-	],
-});
-
-const maskParitySplit = categoryExpansion({
-	category: "split",
-	rows: [
-		[
-			"cross-split",
-			"Cross Split",
-			"十字分割",
-			"texture",
-			"texture-mask",
-			0.55,
-			{ maskShape: "cross", tags: ["cross"], latest: true },
-		],
-		[
-			"triptych-open",
-			"Triptych Open",
-			"三屏拉开",
-			"texture",
-			"texture-mask",
-			0.6,
-			{ maskShape: "triptych", tags: ["panels"] },
-		],
-	],
-});
-
-const maskParityEmoji = categoryExpansion({
-	category: "emoji",
-	rows: [
-		[
-			"heart-expand",
-			"Heart Expand",
-			"爱心扩散",
-			"texture",
-			"texture-mask",
-			0.65,
-			{ maskShape: "heart", tags: ["heart", "love"], popular: true },
-		],
-	],
-});
-
-const maskParityVariety = categoryExpansion({
-	category: "variety",
-	rows: [
-		[
-			"star-expand",
-			"Star Expand",
-			"星形扩散",
-			"texture",
-			"texture-mask",
-			0.6,
-			{ maskShape: "star", tags: ["star"] },
-		],
-		[
-			"gold-dust-dissolve",
-			"Gold Dust Dissolve",
-			"金粉消散",
-			"particle",
-			"particle-dissolve",
-			0.75,
-			{
-				tuning: { intensity: 0.8, frequency: 5, tint: "#ffd76a" },
-				tags: ["gold"],
-			},
-		],
-	],
-});
-
-const maskParityNatural = categoryExpansion({
-	category: "natural",
-	rows: [
-		[
-			"ink-bleed",
-			"Ink Bleed",
-			"水墨晕染",
-			"texture",
-			"texture-mask",
-			0.8,
-			{ maskShape: "ink", tags: ["ink", "水墨"], popular: true },
-		],
-		[
-			"cloud-drift-wipe",
-			"Cloud Drift",
-			"云朵漫卷",
-			"texture",
-			"texture-mask",
-			0.85,
-			{ maskShape: "cloud", tags: ["cloud", "云"] },
-		],
-		[
-			"fog-roll",
-			"Fog Roll",
-			"雾漫",
-			"texture",
-			"texture-mask",
-			0.9,
-			{ maskShape: "fog", tags: ["fog", "雾"] },
-		],
-		[
-			"water-drip-bleed",
-			"Water Drip",
-			"水滴晕染",
-			"texture",
-			"texture-mask",
-			0.75,
-			{ maskShape: "drip", tags: ["drip", "water"], latest: true },
-		],
-		[
-			"snow-mist-dissolve",
-			"Snow Mist Dissolve",
-			"雪雾消散",
-			"particle",
-			"particle-dissolve",
-			0.8,
-			{
-				tuning: { intensity: 0.6, frequency: 3, tint: "#eef4ff" },
-				tags: ["snow"],
-			},
-		],
-	],
-});
-
-const distortionParity = categoryExpansion({
-	category: "distortion",
-	rows: [
-		[
-			"shockwave-cut",
-			"Shockwave Cut",
-			"冲击波切换",
-			"ripple",
-			"shockwave",
-			0.55,
-			{
-				tuning: { intensity: 1, frequency: 1 },
-				tags: ["shockwave"],
-				popular: true,
-			},
-		],
-		[
-			"shockwave-heavy",
-			"Shockwave II",
-			"冲击波Ⅱ",
-			"ripple",
-			"shockwave",
-			0.5,
-			{ tuning: { intensity: 1.6, frequency: 0.7 }, tags: ["impact"] },
-		],
-		[
-			"vortex-twist",
-			"Vortex Twist",
-			"涡旋扭转",
-			"ripple",
-			"vortex",
-			0.65,
-			{ tuning: { intensity: 1 }, tags: ["vortex", "swirl"], popular: true },
-		],
-		[
-			"vortex-gentle",
-			"Vortex Soft",
-			"涡旋轻旋",
-			"ripple",
-			"vortex",
-			0.75,
-			{ tuning: { intensity: 0.5 }, tags: ["swirl", "soft"] },
-		],
-		[
-			"water-pierce",
-			"Water Pierce",
-			"水波穿越",
-			"ripple",
-			"water-ripple",
-			0.6,
-			{
-				tuning: { intensity: 1.3, frequency: 1.4 },
-				tags: ["water"],
-				latest: true,
-			},
-		],
-		[
-			"ripple-touch",
-			"Ripple Touch",
-			"涟漪轻探",
-			"ripple",
-			"water-ripple",
-			0.8,
-			{ tuning: { intensity: 0.5, frequency: 0.8 }, tags: ["gentle"] },
-		],
-		[
-			"fisheye-pierce",
-			"Fisheye Pierce",
-			"鱼眼穿越Ⅲ",
-			"zoom",
-			"zoom-blur",
-			0.5,
-			{ tuning: { intensity: 1.7 }, tags: ["fisheye", "punch"] },
-		],
-		[
-			"glass-shock",
-			"Glass Shock",
-			"玻璃冲击",
-			"glass",
-			"glass-refraction",
-			0.5,
-			{ tuning: { intensity: 1.25, frequency: 1.2 }, tags: ["glass"] },
-		],
-	],
-});
-
-const pseudo3dSlideshow = categoryExpansion({
-	category: "slideshow",
-	rows: [
-		[
-			"cube-rotate",
-			"Cube Rotate",
-			"立方旋转",
-			"page",
-			"cube",
-			0.7,
-			{ tuning: { intensity: 1 }, tags: ["cube", "3d"], popular: true },
-		],
-		[
-			"cube-rotate-2",
-			"Cube Rotate II",
-			"立方旋转Ⅱ",
-			"page",
-			"cube",
-			0.55,
-			{ tuning: { intensity: 1.5 }, tags: ["cube", "fast"] },
-		],
-		[
-			"curtain-open",
-			"Curtain Open",
-			"开幕",
-			"texture",
-			"texture-mask",
-			0.7,
-			{ maskShape: "curtain", tags: ["curtain", "reveal"], popular: true },
-		],
-		[
-			"curtain-open-slow",
-			"Curtain Open II",
-			"开幕Ⅱ",
-			"texture",
-			"texture-mask",
-			1,
-			{ maskShape: "curtain", tags: ["curtain", "cinematic"] },
-		],
-		[
-			"page-flip-up-2",
-			"Page Flip Up II",
-			"翻页Ⅱ",
-			"page",
-			"page-flip",
-			0.6,
-			{ direction: "up", tuning: { intensity: 0.9 }, tags: ["page"] },
-		],
-		[
-			"fold-unfold",
-			"Fold Unfold",
-			"对折展开",
-			"page",
-			"page-flip",
-			0.65,
-			{
-				direction: "left",
-				tuning: { intensity: 1.4 },
-				tags: ["fold"],
-				latest: true,
-			},
-		],
-	],
-});
-
-const parityRoundTwo = categoryExpansion({
-	category: "dissolve",
-	rows: [
-		[
-			"shadow-diagonal-sweep",
-			"Shadow Diagonal Sweep",
-			"阴影斜扫",
-			"texture",
-			"texture-mask",
-			0.6,
-			{ maskShape: "diagonal", tags: ["shadow", "diagonal"], popular: true },
-		],
-		[
-			"push-zoom-defocus",
-			"Push Zoom Defocus",
-			"推镜虚化",
-			"zoom",
-			"zoom-blur",
-			0.6,
-			{ tuning: { intensity: 0.75 }, tags: ["defocus", "push"] },
-		],
-	],
-});
-
-const parityRoundTwoCamera = categoryExpansion({
-	category: "camera",
-	rows: [
-		[
-			"punch-through-3",
-			"Punch Through III",
-			"穿越Ⅲ",
-			"zoom",
-			"zoom-blur",
-			0.4,
-			{ tuning: { intensity: 1.9 }, tags: ["punch", "through"], popular: true },
-		],
-		[
-			"shake-cut-2",
-			"Shake Cut II",
-			"震动Ⅱ",
-			"shake",
-			"shake",
-			0.42,
-			{ tuning: { intensity: 1.2, frequency: 9 }, tags: ["shake"] },
-		],
-		[
-			"memory-pull-2",
-			"Memory Pull II",
-			"回忆拉屏Ⅱ",
-			"whip",
-			"whip-pan",
-			0.75,
-			{
-				direction: "up",
-				tuning: { intensity: 1, tint: "#ffffff" },
-				tags: ["memory"],
-			},
-		],
-		[
-			"screen-pull-left",
-			"Screen Pull Left",
-			"向左拉屏",
-			"whip",
-			"whip-pan",
-			0.5,
-			{
-				direction: "left",
-				tuning: { intensity: 0.85, tint: "#ffffff" },
-				tags: ["pull"],
-				latest: true,
-			},
-		],
-	],
-});
-
-const parityRoundTwoSlideshow = categoryExpansion({
-	category: "slideshow",
-	rows: [
-		[
-			"circle-spin",
-			"Circle Spin",
-			"圆圈旋转",
-			"texture",
-			"texture-mask",
-			0.5,
-			{ maskShape: "clock", tags: ["circle", "spin"] },
-		],
-	],
-});
-
-const parityRoundTwoNatural = categoryExpansion({
-	category: "natural",
-	rows: [
-		[
-			"cloud-drift-2",
-			"Cloud Drift II",
-			"云朵Ⅱ",
-			"texture",
-			"texture-mask",
-			1.1,
-			{ maskShape: "cloud", tags: ["cloud", "slow"] },
-		],
-	],
-});
-
-const parityRoundTwoShooting = categoryExpansion({
-	category: "shooting",
-	rows: [
-		[
-			"exposure-pop-2",
-			"Exposure Pop II",
-			"曝光闪切Ⅱ",
-			"flash",
-			"flash",
-			0.32,
-			{ tuning: { intensity: 1.2, tint: "#ffffff" }, tags: ["exposure"] },
-		],
-		[
-			"zoom-shutter",
-			"Zoom Shutter",
-			"变焦快门",
-			"zoom",
-			"zoom-blur",
-			0.36,
-			{ tuning: { intensity: 0.9 }, tags: ["shutter", "zoom"] },
-		],
-	],
-});
-
-const parityRoundTwoBlur = categoryExpansion({
-	category: "blur",
-	rows: [
-		[
-			"depth-defocus",
-			"Depth Defocus",
-			"纵深虚化",
-			"motion-blur",
-			"motion-blur",
-			0.55,
-			{ direction: "up", tuning: { intensity: 0.5 }, tags: ["depth"] },
-		],
-	],
-});
-
-const mgParity = categoryExpansion({
-	category: "mg",
-	rows: [
-		[
-			"color-swipe-left",
-			"Color Swipe Left",
-			"色块左扫",
-			"wipe",
-			"color-swipe",
-			0.5,
-			{
-				direction: "left",
-				tuning: { tint: "#ffd233" },
-				tags: ["mg", "color"],
-				popular: true,
-			},
-		],
-		[
-			"color-swipe-right",
-			"Color Swipe Right",
-			"色块右扫",
-			"wipe",
-			"color-swipe",
-			0.5,
-			{
-				direction: "right",
-				tuning: { tint: "#4f8bff" },
-				tags: ["mg", "color"],
-			},
-		],
-		[
-			"color-swipe-up",
-			"Color Swipe Up",
-			"色块上扫",
-			"wipe",
-			"color-swipe",
-			0.45,
-			{ direction: "up", tuning: { tint: "#ff5f8f" }, tags: ["mg", "color"] },
-		],
-		[
-			"color-swipe-dark",
-			"Ink Panel Swipe",
-			"墨色幕扫",
-			"wipe",
-			"color-swipe",
-			0.55,
-			{
-				direction: "down",
-				tuning: { tint: "#181818" },
-				tags: ["mg", "dark"],
-				latest: true,
-			},
-		],
-		[
-			"circle-dash",
-			"Circle Dash",
-			"圆形穿梭",
-			"texture",
-			"texture-mask",
-			0.45,
-			{ maskShape: "circle", tags: ["mg", "circle"] },
-		],
-		[
-			"star-pop-in",
-			"Star Pop In",
-			"星形弹入",
-			"texture",
-			"texture-mask",
-			0.5,
-			{ maskShape: "star", tags: ["mg", "star"] },
-		],
-		[
-			"arrow-dash",
-			"Arrow Dash",
-			"箭头冲屏",
-			"texture",
-			"texture-mask",
-			0.4,
-			{ maskShape: "arrow", tags: ["mg", "arrow"] },
-		],
-	],
-});
-
-const emojiParity = categoryExpansion({
-	category: "emoji",
-	rows: [
-		[
-			"heart-flash",
-			"Heart Flash",
-			"爱心闪现",
-			"texture",
-			"texture-mask",
-			0.5,
-			{ maskShape: "heart", tags: ["emoji", "heart"], popular: true },
-		],
-		[
-			"star-mood",
-			"Star Mood",
-			"星星心情",
-			"texture",
-			"texture-mask",
-			0.55,
-			{ maskShape: "star", tags: ["emoji", "star"] },
-		],
-	],
-});
-
-const shootingRoundThree = categoryExpansion({
-	category: "shooting",
-	rows: [
-		[
-			"focus-breath",
-			"Focus Breath",
-			"对焦呼吸",
-			"zoom",
-			"zoom-blur",
-			0.7,
-			{ tuning: { intensity: 0.35 }, tags: ["focus", "breath"] },
-		],
-	],
-});
-
-const fillerSweepDissolve = categoryExpansion({
-	category: "dissolve",
-	rows: [
-		[
-			"slow-cinema-dissolve",
-			"Slow Cinema Dissolve",
-			"慢速叠化",
-			"dissolve",
-			"dissolve",
-			1.5,
-			{ tuning: { intensity: 0.9 }, tags: ["slow", "cinematic"] },
-		],
-	],
-});
-
-const fillerSweepShooting = categoryExpansion({
-	category: "shooting",
-	rows: [
-		[
-			"film-burn-flare",
-			"Film Burn Flare",
-			"胶片烧灼",
-			"light",
-			"light-leak",
-			0.6,
-			{ tuning: { intensity: 1.1, tint: "#ff8c42" }, tags: ["film", "burn"] },
-		],
-		[
-			"slow-shutter-smear",
-			"Slow Shutter Smear",
-			"慢门拖影",
-			"motion-blur",
-			"motion-blur",
-			0.8,
-			{ direction: "right", tuning: { intensity: 0.85 }, tags: ["shutter"] },
-		],
-		[
-			"snap-freeze-flash",
-			"Snap Freeze Flash",
-			"闪拍定格",
-			"flash",
-			"flash",
-			0.34,
-			{
-				tuning: { intensity: 1.1, frequency: 2, tint: "#ffffff" },
-				tags: ["snap"],
-			},
-		],
-	],
-});
-
-const fillerSweepDistortion = categoryExpansion({
-	category: "distortion",
-	rows: [
-		[
-			"vortex-storm",
-			"Vortex Storm",
-			"涡旋强旋",
-			"ripple",
-			"vortex",
-			0.55,
-			{ tuning: { intensity: 1.6 }, tags: ["vortex", "storm"] },
-		],
-	],
-});
-
-const fillerSweepVariety = categoryExpansion({
-	category: "variety",
-	rows: [
-		[
-			"variety-color-swipe",
-			"Variety Color Swipe",
-			"综艺色扫",
-			"wipe",
-			"color-swipe",
-			0.4,
-			{ direction: "right", tuning: { tint: "#ff5f8f" }, tags: ["variety"] },
-		],
-	],
-});
-
-const fillerSweepEmoji = categoryExpansion({
-	category: "emoji",
-	rows: [
-		[
-			"heart-multi-flash",
-			"Heart Multi Flash",
-			"爱心连闪",
-			"flash",
-			"flash",
-			0.4,
-			{
-				tuning: { intensity: 0.85, frequency: 3, tint: "#ff9fbd" },
-				tags: ["heart"],
-			},
-		],
-		[
-			"star-spin-wipe",
-			"Star Spin Wipe",
-			"星星旋转",
-			"texture",
-			"texture-mask",
-			0.45,
-			{ maskShape: "clock", tags: ["star", "spin"] },
-		],
-		[
-			"heartbeat-zoom",
-			"Heartbeat Zoom",
-			"心跳缩放",
-			"zoom",
-			"zoom-blur",
-			0.4,
-			{ tuning: { intensity: 0.5 }, tags: ["heartbeat"] },
-		],
-		[
-			"rainbow-swipe",
-			"Rainbow Swipe",
-			"彩虹色扫",
-			"wipe",
-			"color-swipe",
-			0.45,
-			{ direction: "up", tuning: { tint: "#7c5cff" }, tags: ["rainbow"] },
-		],
-		[
-			"candy-bounce",
-			"Candy Bounce",
-			"糖果弹跳",
-			"shake",
-			"shake",
-			0.5,
-			{ tuning: { intensity: 0.6, frequency: 3 }, tags: ["candy"] },
-		],
-		[
-			"bubble-dissolve",
-			"Bubble Dissolve",
-			"泡泡消散",
-			"particle",
-			"particle-dissolve",
-			0.65,
-			{
-				tuning: { intensity: 0.7, frequency: 4, tint: "#bfe7ff" },
-				tags: ["bubble"],
-			},
-		],
-	],
-});
-
 export const JIANYING_PARITY_TRANSITION_EXPANSIONS: TransitionPreset[] = [
 	...dissolveParity,
 	...cameraParity,
@@ -1157,26 +428,6 @@ export const JIANYING_PARITY_TRANSITION_EXPANSIONS: TransitionPreset[] = [
 	...splitParity,
 	...naturalParity,
 	...shootingParity,
-	...maskParityDissolve,
-	...maskParitySlideshow,
-	...maskParitySplit,
-	...maskParityEmoji,
-	...maskParityVariety,
-	...maskParityNatural,
-	...distortionParity,
-	...pseudo3dSlideshow,
-	...parityRoundTwo,
-	...parityRoundTwoCamera,
-	...parityRoundTwoSlideshow,
-	...parityRoundTwoNatural,
-	...parityRoundTwoShooting,
-	...parityRoundTwoBlur,
-	...mgParity,
-	...emojiParity,
-	...shootingRoundThree,
-	...fillerSweepDissolve,
-	...fillerSweepShooting,
-	...fillerSweepDistortion,
-	...fillerSweepVariety,
-	...fillerSweepEmoji,
+	...JIANYING_MASK_PARITY_EXPANSIONS,
+	...JIANYING_EXTRA_PARITY_EXPANSIONS,
 ];
